Created the UserService class. (#10)
This class takes in user input, validates it, and then utilizes the
accompanying UserRepository to query or modify the database.

Additionally:
Added a new query in UserRepository, findUserByEmail, that
returns a specific user for a given email, if they exist.
Added a new query in UserRepository, findUserByEmailAndPassword, to verify
user login details.
Added a new query in UserRepository, deleteUserByEmail, to facilitate
the deletion of accounts.

Changed the role attribute in the User entity class to use
enums instead of strings.

Reintroduced userID so that users may modify the email associated with their account.